feat(break): Changed authentication to use JWT through Auth0.
This includes migrating all `user_id` columns from integer to text, as the new ids from the tokens a prepended with a provider.
There is still some work left to be done around the `UserService` as that is not really applicable anymore now that several different providers can be used. The frontend should instead talk directly to Github/Auth0 and get the info from there